WIP Uniform power

+++ src/flux_pwr_manager/node_capabilities.h
@@ -2,12 +2,14 @@
 #define FLUX_NODE_CAPABILIITIES_H
 #include "device_type.h"
 #include <stdbool.h>
+#include <unistd.h>
+
 typedef struct {
   int count;
   device_type type;
   bool powercap_allowed;
-  int min_power;
-  int max_power;
+  double min_power;
+  double max_power;
 } device_capability;
 
 typedef struct {
